requirements.txt
gradio>=4.19.2
PyMuPDF>=1.23.26
langchain>=0.1.9
langchain-core>=0.1.27
langgraph>=0.0.26
python-dotenv>=1.0.1
openai>=1.12.0
pydantic>=2.6.3
typing-extensions>=4.10.0
langchain-openai>=0.1.10